nimbix-admin/wrapper-service/requirements.txt
pyyaml
requests
flask
pysftp
